+++ theme/imagecaption/theme.scss
